[platform_device_loader] isolate pki related operations
